starting the pathfinding
